avoid compiler warning (comparison signed/unsigned)
SVN-Revision: 35546

+++ contrib/brl/bseg/boxm2_multi/exe/boxm2_multi_update.cxx
@@ -86,7 +86,7 @@ int main(int argc,  char** argv)
 
   //shuffle frames
   int* frames = new int[cams.size()];
-  for (int i=0; i<cams.size(); ++i) frames[i] = i;
+  for (unsigned int i=0; i<cams.size(); ++i) frames[i] = (int)i;
   boxm2_util::random_permutation(frames, cams.size());
 
   //timing vars
